Refactor Chapter 4 & 5 after Review

+++ src/chapter05/exercice/BasicStopwatchV2.java
@@ -15,7 +15,7 @@ public class BasicStopwatchV2 {
 
     }
 
-    public void launch(int durationInMin) throws InterruptedException {
+    public void launch(double durationInMin) throws InterruptedException {
 
         System.out.println(this.nameStopWatch + " :");
